requirements/build.txt
pytest

# Package requirements
jinja2

# Optional
pandas
matplotlib
Pillow